Some changes for iOS8

+++ src/LiveSDK/Library/Internal/Helpers/StreamReader.m
@@ -69,7 +69,7 @@ const NSUInteger BUFFERSIZE = 4096;
             }
             
             uint8_t buf[BUFFERSIZE];
-            unsigned int len = 0;
+            NSUInteger len = 0;
             len = [_stream read:buf maxLength:BUFFERSIZE];
             if(len) 
             {
